Moving util-like functions to util files or folders
