+++ libs/orchestrator-ui-components/src/index.ts
@@ -1,9 +1,12 @@
 export * from './lib/components/custom-button/custom-button';
 export * from './lib/components/OrchestratorPageTemplate/OrchestratorPageTemplate';
 export * from './lib/components/Table/Table';
+export * from './lib/components/Table/columns';
 
 export * from './lib/theme/defaultOrchestratorTheme';
-export * from './lib/utils';
+
+export * from './lib/utils/getTypedFieldFromObject';
+export * from './lib/utils/getStatusBadgeColor';
 
 export * from './lib/contexts/OrchestratorConfigContext';
