+++ libs/orchestrator-ui-components/src/lib/utils/getStatusBadgeColor.ts
